requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
sounddevice==0.4.7
soundfile==0.12.1
faster-whisper==1.0.3